The app

+++ src/main/java/ru/itmo/p3214/s312198/web/model/PointCheckerRequest.java
@@ -1,24 +1,24 @@
 package ru.itmo.p3214.s312198.web.model;
 
 public class PointCheckerRequest {
-    private Integer x;
+    private Float x;
     private Float y;
     private Integer r;
 
     public PointCheckerRequest() {
     }
 
-    public PointCheckerRequest(Integer x, Float y, Integer r) {
+    public PointCheckerRequest(Float x, Float y, Integer r) {
         this.x = x;
         this.y = y;
         this.r = r;
     }
 
-    public Integer getX() {
+    public Float getX() {
         return x;
     }
 
-    public void setX(Integer x) {
+    public void setX(Float x) {
         this.x = x;
     }
